requirements.txt
aiogram
python-dotenv
requests
beautifulsoup4
rasterio
deep-translator
python-docx
